+++ core/attack/pgdl1.py
@@ -85,7 +85,7 @@ class PGDl1(BaseAttack):
             self.lambda_ = min_lambda_
         else:
             self.lambda_ = max_lambda_
-        adv_x = x.detach().clone().to(torch.double)
+        adv_x = x.detach().clone()
         while self.lambda_ <= max_lambda_:
             _, done = self.get_loss(model, adv_x, label, self.lambda_)
             if torch.all(done):
@@ -129,7 +129,7 @@ class PGDl1(BaseAttack):
         # 4. look for important position
         absolute_grad = torch.abs(gradients).reshape(features.shape[0], -1)
         _, position = torch.max(absolute_grad, dim=-1)
-        perturbations = F.one_hot(position, num_classes=absolute_grad.shape[-1]).double()
+        perturbations = F.one_hot(position, num_classes=absolute_grad.shape[-1])
         perturbations = perturbations.reshape(features.shape)
         directions = torch.sign(gradients) * (perturbations > 1e-6)
