+++ examples/bga_test.py
@@ -97,7 +97,7 @@ def _main():
                                        name=args.model_name,
                                        **hp_params
                                        )
-    model = model.to(dv).double()
+    model = model.to(dv)
     if args.model == 'md_at_pgd':
         at_wrapper = PGDAdvTraining(model)
         at_wrapper.load()
@@ -122,7 +122,7 @@ def _main():
                                      name=args.model_name,
                                      **hp_params
                                      )
-        model = model.to(dv).double()
+        model = model.to(dv)
         model.load()
     elif args.model == 'amd_dnn_plus':
         model = AMalwareDetectionDNNPlus(md_nn_model=None,
@@ -132,7 +132,7 @@ def _main():
                                          name=args.model_name,
                                          **hp_params
                                          )
-        model = model.to(dv).double()
+        model = model.to(dv)
         model.load()
     elif args.model == 'amd_pad_ma':
         adv_model = AMalwareDetectionPAD(model)
@@ -153,7 +153,7 @@ def _main():
     x_mod_integrated = []
     model.eval()
     for x, y in mal_test_dataset_producer:
-        x, y = utils.to_tensor(x.double(), y.long(), model.device)
+        x, y = utils.to_tensor(x, y.long(), model.device)
         adv_x_batch = attack.perturb(model, x, y,
                                      args.steps,
                                      min_lambda_=1e-5,
