+++ examples/gdkdel1_test.py
@@ -111,7 +111,7 @@ def _main():
                                        name=args.model_name,
                                        **hp_params
                                        )
-    model = model.to(dv).double()
+    model = model.to(dv)
     if args.model == 'md_at_pgd':
         at_wrapper = PGDAdvTraining(model)
         at_wrapper.load()
@@ -136,7 +136,7 @@ def _main():
                                      name=args.model_name,
                                      **hp_params
                                      )
-        model = model.to(dv).double()
+        model = model.to(dv)
         model.load()
     elif args.model == 'amd_dnn_plus':
         model = AMalwareDetectionDNNPlus(md_nn_model=None,
@@ -146,7 +146,7 @@ def _main():
                                          name=args.model_name,
                                          **hp_params
                                          )
-        model = model.to(dv).double()
+        model = model.to(dv)
         model.load()
     elif args.model == 'amd_pad_ma':
         adv_model = AMalwareDetectionPAD(model)
@@ -161,7 +161,7 @@ def _main():
     with torch.no_grad():
         c = args.n_center if args.n_center < ben_count else ben_count
         for ben_x, ben_y in ben_test_dataset_producer:
-            ben_x, ben_y = utils.to_tensor(ben_x.double(), ben_y.long(), device=dv)
+            ben_x, ben_y = utils.to_tensor(ben_x, ben_y.long(), device=dv)
             ben_feature_vectors.append(ben_x)
             if len(ben_feature_vectors) * hp_params['batch_size'] >= c:
                 break
@@ -180,7 +180,7 @@ def _main():
     x_mod_integrated = []
     model.eval()
     for x, y in mal_test_dataset_producer:
-        x, y = utils.to_tensor(x.double(), y.long(), model.device)
+        x, y = utils.to_tensor(x, y.long(), model.device)
         adv_x_batch = attack.perturb(model, x, y,
                                      args.steps,
                                      min_lambda_=1e-5,
